pyproject.toml
[tool.poetry]
name = "ai-coworker"
version = "0.1.0"
description = ""
authors = ["Hungjoon Kim <wintersalmon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.8"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.7"
langchain-core = "^0.2.19"
langgraph = "^0.1.8"
duckduckgo-search = "^6.2.0"
langchain-experimental = "^0.0.62"
pygame = "^2.6.0"
json-repair = "^0.28.3"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
